+++ Projektarbeit/src/Insertionsort.java
@@ -5,7 +5,7 @@
  * @author Hermann Witte
  */
 public class Insertionsort implements Sortieralgorithmus {
-    private int vergleiche = 0;
+    private long vergleiche = 0;
     private double zeit;
     private int schreibzugriffe = 0;
     private int length;
@@ -54,7 +54,7 @@ public class Insertionsort implements Sortieralgorithmus {
      * @return Gibt die Anzahl der vergleiche, welcher der Algorithmus benötigt zurück.
      */
     @Override
-    public int getVergleiche() {
+    public long getVergleiche() {
         return vergleiche;
     }
